+++ mm2src/adex_cli/src/main.rs
@@ -5,6 +5,7 @@
 #[cfg(not(target_arch = "wasm32"))] mod cli;
 #[cfg(not(target_arch = "wasm32"))] mod helpers;
 mod logging;
+#[cfg(not(target_arch = "wasm32"))] mod rpc_data;
 #[cfg(not(target_arch = "wasm32"))] mod scenarios;
 #[cfg(all(not(target_arch = "wasm32"), test))] mod tests;
 #[cfg(not(target_arch = "wasm32"))] mod transport;
