+++ mm2src/mm2_rpc/src/data/legacy/activation.rs
@@ -1,0 +1,26 @@
+#[path = "activation/eth.rs"] pub mod eth;
+#[path = "activation/utxo.rs"] pub mod utxo;
+
+use common::serde_derive::{Deserialize, Serialize};
+use mm2_number::BigDecimal;
+
+#[derive(Serialize, Deserialize)]
+pub struct EnabledCoin {
+    pub ticker: String,
+    pub address: String,
+}
+
+pub type GetEnabledResponse = Vec<EnabledCoin>;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct CoinInitResponse {
+    pub result: String,
+    pub address: String,
+    pub balance: BigDecimal,
+    pub unspendable_balance: BigDecimal,
+    pub coin: String,
+    pub required_confirmations: u64,
+    pub requires_notarization: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub mature_confirmations: Option<u32>,
+}
